app/requirements.txt
Flask
mysql-connector
simplejson
searchtweets
tweepy
datetime
bootstrap-flask
flask-sqlalchemy
mysqlclient